Improve selection of element ranges
This CL adds a new method to XmlContext for looking
up a suitable error range for a problem for an Element.
We don't want to select the entire range when it's
a multi-line element; in that case we want to fall
back to the element tag. And if an element has some
other attribute which is more representative, such
as a name attribute, use that one.

Test: Unit tests updated
Change-Id: I3ffb0766a343f9b36500a9826b61bd8c24b91329