Workaround the platform bug for startup CPU profiling
An issue in the android platform was freezing stop profiling that
was exposed by startup CPU Profiling. The cause is that stop profiling
relies on the file close event, however on pre-P devices the file close
doesn't happen until the next garbage collection,
which is undeterministic.
Unfortunately, the only option for pre-P devices is working around
the issue. So, this CL monitors file size instead of close event.

Bug: 74765651
Test: manually
Change-Id: Icef8f331b769efbf33140e92e2dc6df146eed94b