Use javacTask bootclasspath instead of system property
Property sun.boot.class.path has been removed in JDK9.
When running Desugar on JDK9, build would fail as the
task tries to get the value of this property to set up
Desugar --bootclasspath_entry options.

This CL is using the bootclasspath value of the javac
task that was used to compile the code. It will contain
all necessary types, and it is more precise.

BUG: 72872257
Test: existing
Change-Id: Ibc3b32b2982966adcbff075375557c6b302ec043