AndroidX support: Use real AndroidX dep. to test Jetifier.
Since AndroidX dependencies are now available in GMaven and prebuilts,
we should use those real dependencies for testing, instead of fake ones
as before.

Note that Jetifier maps old dependencies to *stable* versions of
AndroidX; however, currently the ones in GMaven and prebuilts are only
*preview* versions, so we need to have a workaround in JetifyTransform
(marked with TODO) that converts stable versions to preview ones.

Test: JetifierTest
Bug: 73306108
Fixes: 77634021
Change-Id: I0d38bab641c1d13a2e09329bbddd845b0c51c50e