[AVD Mgr] Manage named snapshots
The Emulator allows the user to snapshot the device state. The
user can then rename and annotate those snapshots. The user can
select one of these snapshots for use when the AVD next launches.

This CL allows the user to access those controls from Studio. For
the simplest operations, the control is directly within Studio.
For more complex operations, Studio invokes the Emulator to display
just the UI page that deals with managing snapshots.

This feature is contingent on Emulator support, which is first
available in Emulator version 27.2.5.

Bug: 76205895
Test: Added ConfigureAvdOptionsStepTest.testPopulateSnapshotList
Change-Id: I4eca10c12910b4d21f5f32345b6f1cc8c76aad99