Fix JVMTI agent initialization order
With CPU startup profiling, the heartbeat thread now starts
earlier than when the MemoryTrackingEnv finishes iniitalization.
When Studio Profilers receives the heartbeat, it attempts to
start live allocation tracking which would fail as the memory
portion of the agent is not fully ready yet. This CL switches
the order back to the way it was before.

Bug: 77875260
Test: N/A
Change-Id: I27cff0821575f4988a6c9f59d86642861786baa0