[New Dynamic Module] Add Activities to new Module
When adding a new Activity to a Dynamic Module, all resources
referenced in the Manifest must be copied to the Base Feature.
This is a limitation of the Build System that requires that all
the values of the merged manifest are in the Base. This is also
similar to Instant Apps.

Bug: 78512556
Test-Strategy: ui.test.groups=PROJECT_WIZARD
Test: All tests in PROJECT_WIZARD group
Change-Id: Ic2f86d159278dd57903dbdf6f1b40416a141fff9