Publish consumer proguard files to base module
This CL is the 7th in a series to provide proguard/R8
support for multi-APK projects.

This CL publishes consumer proguard files from the feature
modules to the base modules when shrinking is enabled in
a multi-apk project.

This CL also merges the feature modules' runtime dependencies'
classes with the feature's, and publishes the merged classes
and java resources for the base module's transform pipeline.

Test: MinifyFeaturesTest, MergeClassesTransformTest
Bug: 74433581
Change-Id: I0d0dbcd3af5136c314a16cd0f9f3d5447692fb5c