Introduces system property for CPU ART stop timeout
Introduces a system property "profiler.cpu.art.stop.timeoutSec"
for defining the wait timeout for ART to stop the ongoing profiling
and close the trace file. For usual use cases the default value of
the parameter (5 Sec) should be enough, however for performance
intensive applications 5 Seconds might be not enough, as the app may
take all CPU resources. Thus, for this use cases the users can increase
the timeout.

Test: manually by adding "profiler.cpu.art.stop.timeoutSec=123" from
      Help > Edit Custom Properties.
Fixes: 78456396
Bug: 77282701

Change-Id: I5c379471640204ee6de8a9d1d5ae149735c293d1