Make data binding work with task output caching (1/2).
(Note: This doesn't work when Kapt is present yet.)

Data binding injects a task called DataBindingExportInfo, which defines
several inputs that affect the results of data binding annotation
processing.

In incremental builds, when these inputs are changed, data binding needs
to be retriggered. So far, we've done this by:
 1. Letting JavaCompile depend on a file generated by
    DataBindingExportInfo task.
 2. Make the contents of the generated file unique per a set of inputs.

At step 2, ideally we should compute the hash of the inputs to include
in the file, but for convenience, we include a random ID per build in
the file. Although this makes incremental builds work, it makes task
output caching of JavaCompile useless since the inputs to JavaCompile
contains that file with a random ID.

To fix this, we need to remove those inputs from the
DataBindingExportInfo task and make JavaCompile depend on them directly
via the AnnotationProcessorOptions.compilerArgumentProvider() API. (The
second part was done in an earlier CL.)

The DataBindingExportInfoTask no longer generates the file with a random
ID but is still needed for a different reason: It will now create an
empty class annotated with @BindingBuildInfo, so that JavaCompile
invokes data binding even when the rest of the source code does not have
data binding annotations.

Bug: 69243050
Test: Added DataBindingCachingTest
      Added tests in DataBindingIncrementalTest
Change-Id: I1b69c0a86ef981f22634862be67d0143760dec2c