Do not use -dump by default with Proguard
By default, -dump should not be used with Proguard.
This option lets ProGuard write out a text version
of the processed bytecode to the build directory,
which can be tens or even hundreds of megabytes
with little value. E.g. for a default project
created in Studio, dump.txt is 13MB.

BUG: 79669448
Test: existing
Change-Id: I33d7f8f8c2fb315baf674d53539a6053ca43d61a