[Plugin-wide change] Always use processed aars/jars.
Aars/jars may need to be processed (e.g., jetified to AndroidX) before
they can be used.

This CL ensures that all places in the plugin that used to access
original aars/jars now need to access processed aars/jars instead. One
or multiple ArtifactTransforms may be triggered to convert original
aars/jars into processed ones on the fly.

The exceptions to this rule include:
  - android.jar: It does not need to be processed (and Jetifier
    currently also fails to process it).
  - lint.jar in GlobalScope#getLocalCustomLintChecks()
  - Jars used by LintStandaloneTask as this task is executed without the
    AGP
  - Aars/jars when building the model, see ArtifactDependencyGraph.java.
    The reason is that the aars/jars published by library modules may
    not yet exist. If we query for processed aars/jars, it will trigger
    ArtifactTransform to run on non-existent files, which will fail.
    I've filed https://issuetracker.google.com/110054209 to track if
    there is a workaround for this.

Currently, Jetifier is the only one that does this transform, but in
theory, there could be multiple others.

The trigger for this change is
https://issuetracker.google.com/78683535, where annotation processors
were not jetified because the JavaCompile task queried for original
jars, not the jetified ones, in the annotation processor configuration.

Fixes: 78683535
Bug: 110054209
Test: Existing tests + Updated JetifierTest
Change-Id: Ia1e5f23c2f1af638af9aee4ba46092278b49704c