Crashlytics: Enforce minimum version of Crashlytics plugin.
In versions prior to 1.25.4, the Crashlytics (io.fabric.tools:gradle)
plugin repackaged ASM without renaming it, resulting in run-time
conflicts with other plugins that use a different version of ASM.

This CL enforces minimum version 1.25.4 when using that plugin.

The issue will be reported as a sync error during sync and a runtime
error when building.

Fixes: 79997489
Test: PluginVersionCheckTest
Change-Id: Ib91ed27e7d0b277203d0a4f8f0d4650e7b92b0b3